구매내역 상세 구현했지만 이미지 무한 요청 에러 발생, 수정중
