+++ src/app/customer/orders/[orderId]/orderDetailClient.tsx
@@ -1,0 +1,223 @@
+// src/app/customer/orders/[orderId]/OrderDetailClient.tsx
+// 이 파일은 클라이언트 컴포넌트입니다.
+'use client';
+
+import { useEffect, useState } from 'react';
+import Image from 'next/image';
+import { OrderResponseDTO } from '@/types/orders/orderResponseDTO'; // DTO 타입 임포트
+import { OrderItemResponseDTO } from '@/types/orders/orderItemResponseDTO'; // DTO 타입 임포트
+
+
+// --- 데이터 페칭 함수 (클라이언트에서 실행) ---
+async function fetchOrderDetail(orderId: number, token: string): Promise<OrderResponseDTO> {
+    const url = `http://localhost:8080/api/customer/orders/${orderId}`;
+    console.log("클라이언트에서 주문 상세 정보를 가져오는 중:", url);
+    console.log("Authorization 헤더에 포함될 토큰 (앞 10자):", token ? `Bearer ${token.substring(0, 10)}...` : "토큰 없음");
+
+    const response = await fetch(url, {
+        cache: "no-store", // 캐시 사용 안 함
+        redirect: 'follow', // 리다이렉트 따름
+        headers: {
+            'Authorization': token ? `Bearer ${token}` : '', // localStorage에서 가져온 토큰 사용
+            'Content-Type': 'application/json'
+        }
+    });
+
+    if (!response.ok) {
+        const errorData = await response.text();
+        console.error(`주문 상세 API 오류: ${response.status} - ${errorData}`);
+        throw new Error(`주문 상세 정보를 불러오는 데 실패했습니다: ${response.statusText || '알 수 없는 오류'}`);
+    }
+
+    const contentType = response.headers.get("content-type");
+    if (!contentType || !contentType.includes("application/json")) {
+        const textResponse = await response.text();
+        console.error("API 응답이 JSON 형식이 아닙니다 (예상치 못한 응답):", textResponse);
+        throw new Error("API 응답이 JSON 형식이 아닙니다.");
+    }
+
+    return response.json(); // JSON 데이터 파싱 및 반환
+}
+
+// ⭐ 수정된 props 인터페이스: orderId를 직접 받습니다.
+interface OrderDetailClientProps {
+    orderId: string;
+}
+
+// --- 주문 상세 페이지 클라이언트 컴포넌트 ---
+export default function OrderDetailClient({ orderId }: OrderDetailClientProps) {
+    const [orderData, setOrderData] = useState<OrderResponseDTO | null>(null);
+    const [error, setError] = useState<string | null>(null);
+    const [loading, setLoading] = useState<boolean>(true);
+
+    useEffect(() => {
+        let token: string | null = null;
+        if (typeof window !== 'undefined') {
+            // ⭐ localStorage에서 'auth-storage' 키로 값을 가져옵니다.
+            const storedStateString = localStorage.getItem('auth-storage');
+            console.log('localStorage에서 가져온 원시 데이터 (auth-storage):', storedStateString); // 디버깅 로그
+
+            if (storedStateString) {
+                try {
+                    const parsedState = JSON.parse(storedStateString);
+                    // {state: {accessToken: ...}, version: 0} 구조이므로 parsedState.state.accessToken 에 접근합니다.
+                    token = parsedState.state?.accessToken;
+                    console.log('JSON.parse 후 추출된 accessToken:', token ? `가져옴 (길이: ${token.length})` : '없음'); // 디버깅 로그
+                } catch (e) {
+                    console.error('localStorage 데이터 파싱 오류:', e);
+                    setError("인증 정보 파싱에 실패했습니다. 다시 로그인 해주세요.");
+                    setLoading(false);
+                    return;
+                }
+            }
+        }
+
+        // 최종적으로 token 변수에 유효한 문자열이 들어 있는지 확인합니다.
+        console.log('useEffect 최종 확인: 토큰 존재 여부:', !!token);
+        if (!token) {
+            setError("로그인 토큰이 없습니다. 다시 로그인 해주세요.");
+            setLoading(false);
+            console.log('⛔ API 호출 중단: 토큰이 없습니다.'); // API 호출이 중단되는 지점
+            return;
+        }
+
+        // string 타입의 orderId를 number로 변환하여 API 호출에 사용합니다.
+        const numericOrderId = Number(orderId);
+        if (isNaN(numericOrderId) || numericOrderId <= 0) {
+            setError("유효하지 않은 주문 ID입니다.");
+            setLoading(false);
+            return;
+        }
+
+        const getOrder = async () => {
+            try {
+                // 변환된 numericOrderId와 토큰을 사용하여 데이터 페칭
+                const data = await fetchOrderDetail(numericOrderId, token as string); // `token`이 null이 아님을 TypeScript에 알림
+                setOrderData(data);
+            } catch (err) {
+                console.error("주문 상세 정보를 가져오는 데 실패:", err);
+                setError(err instanceof Error ? err.message : "주문 상세 정보를 불러오는 중 알 수 없는 오류가 발생했습니다.");
+            } finally {
+                setLoading(false);
+            }
+        };
+
+        getOrder();
+    }, [orderId]); // orderId가 변경될 때마다 useEffect 재실행
+
+    // --- 로딩, 에러, 데이터 없음 상태에 따른 UI 렌더링 ---
+    if (loading) {
+        return (
+            <div className="container mx-auto p-4 text-center font-inter">
+                <p>주문 정보를 불러오는 중...</p>
+            </div>
+        );
+    }
+
+    if (error) {
+        return (
+            <div className="container mx-auto p-4 font-inter">
+                <h1 className="text-3xl font-bold mb-6 text-center text-gray-800">
+                    주문 상세
+                </h1>
+                <div
+                    className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded-lg relative mb-4"
+                    role="alert"
+                >
+                    <strong className="font-bold">오류: </strong>
+                    <span className="block sm:inline">{error}</span>
+                </div>
+            </div>
+        );
+    }
+
+    if (!orderData) {
+        return (
+            <div className="container mx-auto p-4 text-center text-gray-700 font-inter">
+                <h1 className="text-3xl font-bold mb-6">주문 상세</h1>
+                <p>주문 정보를 찾을 수 없습니다.</p>
+            </div>
+        );
+    }
+
+    // --- 주문 상세 정보 UI 렌더링 (기존과 동일) ---
+    return (
+        <div className="container mx-auto p-4 bg-gray-50 min-h-screen font-inter">
+            <h1 className="text-4xl font-extrabold mb-8 text-center text-gray-900 leading-tight">
+                주문 상세 정보
+            </h1>
+
+            <div className="bg-white shadow-xl rounded-lg p-8 mb-8 border border-gray-200">
+                <div className="grid grid-cols-1 md:grid-cols-2 gap-6 text-gray-700">
+                    <div>
+                        <p className="text-lg font-semibold mb-2">주문 번호: <span className="font-normal text-blue-600">{orderData.orderId}</span></p>
+                        <p className="text-lg font-semibold mb-2">고객 ID: <span className="font-normal">{orderData.customerId}</span></p>
+                        <p className="text-lg font-semibold mb-2">배송 주소: <span className="font-normal">{orderData.deliveryAddress}</span></p>
+                        <p className="text-lg font-semibold mb-2">총 주문 가격: <span className="font-normal text-green-700">{orderData.totalPrice.toLocaleString()}원</span></p>
+                    </div>
+                    <div>
+                        <p className="text-lg font-semibold mb-2">주문 일시: <span className="font-normal">{new Date(orderData.orderCreatedAt).toLocaleString()}</span></p>
+                        <p className="text-lg font-semibold mb-2">최종 업데이트: <span className="font-normal">{new Date(orderData.updatedAt).toLocaleString()}</span></p>
+                        <p className="text-lg font-semibold mb-2">결제 방식: <span className="font-normal">{orderData.paymentType}</span></p>
+                        <p className="text-lg font-semibold mb-2">배송비: <span className="font-normal">{orderData.deliveryFee.toLocaleString()}원</span></p>
+                        <p className="text-lg font-semibold mb-2">수령인: <span className="font-normal">{orderData.receiverName}</span></p>
+                        <p className="text-lg font-semibold mb-2">연락처: <span className="font-normal">{orderData.phone}</span></p>
+                    </div>
+                </div>
+                <p className="text-lg font-semibold mt-4">
+                    주문 상태:{" "}
+                    <span
+                        className={`relative inline-block px-4 py-1 font-bold leading-tight rounded-full ${
+                            orderData.orderStatus === "ORDER"
+                                ? "bg-green-100 text-green-800"
+                                : "bg-red-100 text-red-800"
+                        }`}
+                    >
+            {orderData.orderStatus === "ORDER" ? "주문 완료" : orderData.orderStatus === "CANCEL" ? "주문 취소" : orderData.orderStatus}
+          </span>
+                    <span className="ml-4 text-gray-600">({orderData.deliveryStatus})</span>
+                </p>
+            </div>
+
+            <div className="bg-white shadow-xl rounded-lg p-8 border border-gray-200">
+                <h2 className="text-3xl font-bold mb-6 text-gray-800 border-b pb-3">주문 상품 목록</h2>
+                {orderData.orderItems.length === 0 ? (
+                    <p className="text-lg text-gray-600 text-center py-4">주문된 상품이 없습니다.</p>
+                ) : (
+                    <div className="space-y-6">
+                        {orderData.orderItems.map((item) => (
+                            <div key={item.productId} className="flex flex-col md:flex-row items-center space-y-4 md:space-y-0 md:space-x-6 p-4 border border-gray-200 rounded-lg shadow-sm hover:shadow-md transition-shadow duration-200">
+                                <div className="flex-shrink-0">
+                                    {/* 이미지 URL 유효성 검사 강화 */}
+                                    {item.imageUrl && typeof item.imageUrl === 'string' && item.imageUrl.trim() !== '' ? (
+                                        <Image
+                                            src={item.imageUrl}
+                                            alt={item.productName}
+                                            width={96}
+                                            height={96}
+                                            className="rounded-lg object-cover w-24 h-24"
+                                            onError={(e) => {
+                                                e.currentTarget.src = `https://placehold.co/96x96/aabbcc/ffffff?text=No+Image`;
+                                                e.currentTarget.onerror = null;
+                                            }}
+                                        />
+                                    ) : (
+                                        // 이미지 URL이 없거나 유효하지 않을 때만 이 대체 UI를 렌더링합니다.
+                                        <div className="flex items-center justify-center bg-gray-200 rounded-lg w-24 h-24 text-gray-500 text-sm">
+                                            이미지 없음
+                                        </div>
+                                    )}
+                                </div>
+                                <div className="flex-grow text-center md:text-left">
+                                    <h3 className="text-xl font-semibold text-gray-800 mb-1">{item.productName}</h3>
+                                    <p className="text-lg text-gray-600 mb-1">{item.quantity}개</p>
+                                    <p className="text-xl font-bold text-blue-700">{item.price.toLocaleString()}원</p>
+                                </div>
+                            </div>
+                        ))}
+                    </div>
+                )}
+            </div>
+        </div>
+    );
+}
